feat(template): sync templates when syncing substates (#1242)
Description
---
Synchronizing when a VN is just added to the network or syncing new
substates. We try to fetch template binary from its owning committee to
make our chance higher to get it.
There is also a logic put to periodically check for pending templates,
so if we know that we must have a template ready, just try to fetch
again from other VNs.

Motivation and Context
---
When new nodes are joining the network, they did not have the template
binaries as so we needed synchronization.

How Has This Been Tested?
---
1. Start a local swarm with X VNs
2. Publishing a new template (via CLI is simpler)
3. Adding a new VN (or more)
4. Mine to next epoch to make VN(s) available
5. See if new VN(s) have the new template already registered

What process can a PR reviewer use to test or verify this change?
---


Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify

---------

Co-authored-by: Stan Bondi <sdbondi@users.noreply.github.com>