feat(wallet): using correct network (#1249)
Description
---
For now we use the correct network configured and also validating on the
validator node side in mempool.

Motivation and Context
---
Closes #1241 

How Has This Been Tested?
---
Start a fresh swarm locally and send any transaction through wallet
daemon. It is expected to still work.

What process can a PR reviewer use to test or verify this change?
---


Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify