fix: cucumber test for events, remove resource event (#1268)
Description
---
fix: failing cucumber
remove resource events for every vault withdraw/deposit

Motivation and Context
---
When a vault deposit/withdraw occurs we used to emit an event for the
resource. This results in extra data that may not be useful. If we do
need this, we should perhaps make it an option on the resource. For the
time being, it's easier to add something later when we know that we need
it and nothing currently relies on it, than to remove it later and hope
that it doesn't break anything, so it is removed here. It is worth
noting that an indexer can locally infer this event based on the vault
event, thereby still allowing deposits by resource to be searched.

How Has This Been Tested?
---
CI

What process can a PR reviewer use to test or verify this change?
---


Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify