fix(template-manager): fix builtin check for exist with status (#1282)
Description
---
Fixes check for existance on template_exists when passing in a builtin
template with a status

Motivation and Context
---
We validate that templates are not marked as invalid in the
TemplateExists validation. If a builtin template address is passed in,
it always returns true regardless of the queried status. This is not
correct and has been corrected in this PR to return false if the status
is not Active

How Has This Been Tested?
---
Existing tests

What process can a PR reviewer use to test or verify this change?
---
CI

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify